ANT fixes for screenshots and "scene" flag setting (#249)
* Update ANT.py

Upload form expects screenshot URLs to be given as a string with one URL per line. Passing a list results in just the first screenshot being added to the upload.

* Update ANT.py

Up to 4 screenshots are now allowed instead of just 2.

* Update ANT.py

Set the "scene" flag (which actually has the form ID of "censored", presumably as a relic of a past change) if the metadata indicates that this is a scene release.